Add creating setup file for Bash
A setup script that can be manually sourced in the Bash interpreter is
created in the project directory of a model. Its purpose is to activate
the virtual environment, set relevant environment variables, and create
useful aliases. It comprises 2 sections:

- 'Config' - specifies necessary variables used during the actual setup
  process;
- 'Setup'- performs the actual setup process.

Some of the variables are automatically populated with appropriate settings
during an initial setup of a model.