Add main functionality
The main package functionality is added, which is performing an initial
setup of a model. First, the desired configuration is determined based
on a provided configuration file and the system configuration is also
determined, then a virtual environment in the project directory of the
model is created, and finally necessary dependencies are installed in
the virtual environment, followed by an optional creating of additional
directories and files that will be used when running simulations.

Additionally, since performing an initial setup of a model requires that
some necessary information be provided using command line arguments,
parsing command line arguments is also added.