Add command line argument for performing initial setup of a model
An optional command line argument for performing an initial setup of a
model is added, which specifies that third-party package 'virtualenv'
should be used for creating a virtual environment instead of package
'venv' from the standard library.

Additionally, the information displayed when confirming configuration is
augmented to include the name of the package to be used for creating a
virtual environment.